Moved some experiments to deprecated folder
