+++ include/daw/daw_check_exceptions.h
@@ -6,6 +6,7 @@
 // Official repository: https://github.com/beached/header_libraries
 //
 
+
 #pragma once
 
 #include "ciso646.h"
