+++ include/daw/daw_string_view.h
@@ -10,7 +10,7 @@
 #include "daw_string_view_version.h"
 
 #if DAW_STRINGVIEW_VERSION == 1
-#include "daw_string_view1.h"
+#include "deprecated/daw_string_view1.h"
 #else
 #include "daw_string_view2.h"
 #endif
