+++ tests/daw_bounded_string_test.cpp
@@ -7,7 +7,7 @@
 //
 
 #include "daw/daw_benchmark.h"
-#include "daw/daw_bounded_string.h"
+#include "daw/deprecated/daw_bounded_string.h"
 #include "daw/daw_string_view.h"
 #include "daw/daw_string_view_fwd.h"
 #include "daw/daw_utility.h"
