+++ tests/daw_clumpy_sparsy_test.cpp
@@ -7,7 +7,7 @@
 //
 
 #include "daw/daw_benchmark.h"
-#include "daw/daw_clumpy_sparsy.h"
+#include "daw/deprecated/daw_clumpy_sparsy.h"
 
 void clumpy_sparsy_test_001( ) {
 	// TODO: Why are we not testing?
