+++ tests/daw_cstring_test.cpp
@@ -7,7 +7,7 @@
 //
 
 #include "daw/daw_benchmark.h"
-#include "daw/daw_cstring.h"
+#include "daw/deprecated/daw_cstring.h"
 
 void daw_cstring_01( ) {
 	daw::cstring a = "a";
