+++ tests/daw_string_split_range_test.cpp
@@ -7,8 +7,8 @@
 //
 
 #include "daw/daw_benchmark.h"
-#include "daw/daw_string_split_range.h"
-#include "daw/daw_string_view1.h"
+#include "daw/deprecated/daw_string_split_range.h"
+#include "daw/deprecated/daw_string_view1.h"
 
 #include <algorithm>
 #include <array>
